feat: implemented new typed Result object
