--- autograder/core/grading/models/result.py
@@ -1,25 +1,0 @@
-class Result:
-    """
-    Represents the result of an assignment submission.
-    Contains a final score generated by the Scorer class and the Graders.
-    Allows different report generation methods for feedback.
-    """
-    def __init__(self,final_score,author,submission_file,base_results,bonus_results,penalty_results):
-        self.final_score = final_score
-        self.author = author
-        self.submission_file = submission_file
-        self.base_results = base_results
-        self.bonus_results = bonus_results
-        self.penalty_results = penalty_results
-
-    def __str__(self):
-        display = f"Student's usename: {self.author}\n"
-        display += f"Final score: {self.final_score}\n"
-        display += f"Number of base tests passed: {len(self.base_results['passed'])}\n"
-        display += f"Number of base tests failed: {len(self.base_results['failed'])}\n"
-        display += f"Number of bonus tests passed: {len(self.bonus_results['passed'])}\n"
-        display += f"Number of bonus tests failed: {len(self.bonus_results['failed'])}\n"
-        display += f"Number of penalty tests Detected: {len(self.penalty_results['passed'])}\n"
-        return display
-
-
