+++ autograder/core/models/result.py
@@ -1,0 +1,20 @@
+from typing import List
+from test_result import TestResult
+class Result:
+    """
+    Represents the result of an assignment submission.
+    Contains a final score generated by the Scorer class and the Graders.
+    Allows different report generation methods for feedback.
+    """
+    def __init__(self,final_score,author,submission_file,base_results: List[TestResult],bonus_results:List[TestResult],penalty_results:List[TestResult]):
+        self.final_score = final_score
+        self.author = author
+        self.submission_file = submission_file
+        self.base_results = base_results
+        self.bonus_results = bonus_results
+        self.penalty_results = penalty_results
+
+    def __repr__(self):
+        return f"Result(final_score={self.final_score}, author={self.author}, submission_file={self.submission_file}, base_results={self.base_results}, bonus_results={self.bonus_results}, penalty_results={self.penalty_results})"
+
+
